% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adhoc_polyserial.R
\name{adhoc_polyserial}
\alias{adhoc_polyserial}
\title{Adhoc polyserial estimator}
\usage{
adhoc_polyserial(x, y, maxcor = 0.9999, more_verbose = F)
}
\arguments{
\item{x}{a discrete or continuous vector}

\item{y}{a discrete or continuous vector}

\item{maxcor}{maximal allowed correlation so as to make sure that we stay away from the boundary}

\item{more_verbose}{prints some extra information}
}
\value{
a value between [-1,1]
}
\description{
Returns the nonparanormal correlation estimate of a couple of randomvariables
where one is ordinal and the other one is continuous
}
