% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edgenumber.R
\name{edgenumber}
\alias{edgenumber}
\title{Edgenumber}
\usage{
edgenumber(Precision = Precision, cut = 0)
}
\arguments{
\item{Precision}{a square matrix whose entries are equal to negative partial correlations.}

\item{cut}{thresholding parameter when entries are treated to be non-zero.}
}
\value{
a number corresponding to the number of edges present in the underlying graph.
}
\description{
Calculates number of edges given some precision matrix. Thresholding is allowed but defaulted to 0.
}
