% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_hat.R
\name{f.hat}
\alias{f.hat}
\title{Estimate monotone transformation functions}
\usage{
f.hat(x)
}
\arguments{
\item{x}{a numeric vector}
}
\value{
a numeric vector of same length as input
}
\description{
Implements the Winsorized estimator in Liu, Lafferty, Wasserman (2009),
}
\references{
Liu, Han, Lafferty, John and Wasserman, Larry. (2009). The nonparanormal: Semi-parametric estimation of high dimensional undirected graphs.Journal of Machine Learn-ing Research10(80), 2295–2328.
}
