% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hume.R
\docType{package}
\name{hume}
\alias{hume}
\title{hume: A package for high-dimensional mixed graph estimation}
\description{
The hume package provides two important functions:
\code{\link{mixed.graph.gauss}} and \code{\link{mixed.graph.nonpara}}
}
