% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixed_graph_gauss.R
\name{mixed.graph.gauss}
\alias{mixed.graph.gauss}
\title{Mixed Graph Estimation under Gaussian assumption}
\usage{
mixed.graph.gauss(data = data, verbose = T, nlam = 50, param = 0.1, names = F)
}
\arguments{
\item{data}{A data set of dimension \eqn{n \times d}}

\item{verbose}{Prints some useful information, defaults to `TRUE`}

\item{nlam}{Length of the glasso path. Default is 50}

\item{param}{Value to which the additional dimensionality penalty in the eBIC should be set. Default is .1}

\item{names}{Can be either boolean of a character vector of length \eqn{d}. Default is `FALSE`}
}
\value{
A list with the following elements is returned:
\itemize{
  \item Estimated Precision Matrix - The estimated precision matrix. If `partial = TRUE` the entries correspond to partial correlations.
  \item Adjacency Matrix - The corresponding adjacency matrix that encodes the undirected graph.
  \item Sample Correlation Matrix - The sample correlation matrix plugged into the glasso routine.
  \item Edgenumber - The number of edges in the graph
  \item Max Degree - The largest number of connections across all nodes in the graph.
}
}
\description{
Estimates precision matrix and adjacency matrix, i.e. the undirected graph,
given a (potentially high dimensional) data set that contains discrete and continuous variables.
If nominal variables are present, they should be transformed to a dummy system.
}
\examples{
require(huge)
latent.data = huge::huge.generator(n = 50, d = 60, graph = "hub", g = 4)$data
data <- latent.data
binarize <- ncol(latent.data)/2
pbin <- stats::runif(floor(binarize),.4,.6)
for(i in 1:binarize){
   data[,i] <- stats::qbinom(stats::pnorm(scale(latent.data[,i])),size=1,prob = pbin[i])
}
mixed.graph <- mixed.graph.gauss(data)
latent.graph <- mixed.graph.gauss(latent.data)
all.equal(mixed.graph$`Adjacency Matrix`, latent.graph$`Adjacency Matrix`)

}
