% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/npn_pearson.R
\name{npn.pearson}
\alias{npn.pearson}
\title{Nonparanormal product moment correlation}
\usage{
npn.pearson(cont, disc)
}
\arguments{
\item{cont}{numeric vector}

\item{disc}{categorical vector; can be factor or integer}
}
\value{
a correlation, i.e. a value in [,1,1]
}
\description{
Calculates the nonparanormal version of the pearson correlation between a continuous and discrete variable
}
