% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/omega_select.R
\name{omega.select}
\alias{omega.select}
\title{Precision Matrix Selection}
\usage{
omega.select(x = x, param = param, n = n, s = s, partial = T)
}
\arguments{
\item{x}{glasso path contained in a `huge` object}

\item{param}{\eqn{\theta} parameter for additional high-dimensional penalty}

\item{n}{sample size for likelihood recalculation}

\item{s}{sample covariance or correlation matrix that was used to estimate the glasso path}

\item{partial}{if set to `TRUE` the resulting precision matrix entries are negated so as to get partial correlations}
}
\value{
a symmetrix matrix of the same dimension as `s`
}
\description{
For a given glasso path (in the form of a `huge` object) the
function recalculates the unpenalized likelihood via the `glasso()` function
and selects the model, i.e. the graph, with the smallest eBIC
\deqn{    eBIC_\theta = -2 \ell^{(n)}(\hat{\Omega}(E)) + \lvert E \rvert \log(n) + 4 \lvert E \rvert \theta \log(d)}
(see Foygel & Drton (2010)).
By default `partial = TRUE` which converts the entries in the resulting precision
matrix to partial correlations.
}
\references{
Foygel, Rina and Drton, Mathias. (2010).
Extended Bayesian Information Criteria for Gaussian Graphical Models.
In: Lafferty, J., Williams, C., Shawe-Taylor, J., Zemel, R. andCulotta,  A.  (editors),
Advances  in  Neural  Information  Processing  Systems,  Volume  23.Curran Associates, Inc. pp. 604–612.
}
