% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spearman.R
\name{spearman}
\alias{spearman}
\title{Calculate Spearman's rho}
\usage{
spearman(x, y)
}
\arguments{
\item{x}{numeric vector}

\item{y}{numeric vector}
}
\value{
a number in [-1,1]
}
\description{
Calculate Spearman's rho
}
\examples{
x <- sample(10,10)
y <- sin(x)
spearman(x,y)
}
